refactoring(split restaurant page with components)
