+++ src/paragraphs.ts
@@ -1,0 +1,7 @@
+export const PROVIDENCE_FOOD_DELIVERY = `Providence Food Delivery`;
+export const PROVIDENCE_FOOD_DELIVERY_CONTENT = `Have your favorite Providence restaurant food delivered to your door
+            with Uber Eats. Whether you want to order breakfast, lunch, dinner,
+            or a snack, Uber Eats makes it easy to discover new and nearby
+            places to eat in Providence. Browse tons of food delivery options,
+            place your order, and track it by the minute.`;
+export const EXPLORE_BY_CATEGORY = `explore by category`;
